+++ bean-searcher-solon-plugin/src/main/java/cn/zhxu/bs/solon/BeanSearcherProperties.java
@@ -19,7 +19,6 @@ import java.time.ZoneId;
 import java.util.HashMap;
 import java.util.Map;
 
-
 @Inject("${bean-searcher}")
 @Configuration
 public class BeanSearcherProperties {
@@ -111,6 +110,12 @@ public class BeanSearcherProperties {
 		 * @see DefaultParamResolver#setSelectExcludeName(String)
 		 */
 		private String selectExclude = "selectExclude";
+
+		/**
+		 * 参数非法时是否抛出异常
+		 * @since v4.2.3
+		 */
+		private boolean failOnError = false;
 
 		/**
 		 * 过滤器配置
@@ -282,6 +287,14 @@ public class BeanSearcherProperties {
 
 		public Group getGroup() {
 			return group;
+		}
+
+		public boolean isFailOnError() {
+			return failOnError;
+		}
+
+		public void setFailOnError(boolean failOnError) {
+			this.failOnError = failOnError;
 		}
 
 		public Pagination getPagination() {