+++ CHANGELOG.md
@@ -1,8 +1,19 @@
-# v4.2.3 @ Next
+# Next
 
-* 增强 `DateParamConvertor`，支持 `java.util.Date` 子类的转换，并新增 `target` 属性，支持配置转换目标类型
-* 增强 `DateTimeParamConvertor`，支持 `java.util.Date` 子类的转换，并新增 `target` 属性，支持配置转换目标类型
-* 增强 `TimeParamConvertor`，新增 `target` 属性，支持配置转换目标类型
+* Bean Searcher
+  * 增强 `DateParamConvertor`，支持 `java.util.Date` 子类的转换，并新增 `target` 属性，支持配置转换目标类型
+  * 增强 `DateTimeParamConvertor`，支持 `java.util.Date` 子类的转换，并新增 `target` 属性，支持配置转换目标类型
+  * 增强 `TimeParamConvertor`，新增 `target` 属性，支持配置转换目标类型
+* Bean Searcher Boot Starter
+  * 新增配置项 `bean-searcher.params.fail-on-error` 指定参数错误时，是否抛出异常，默认 `false` (TODO)
+  * 新增配置项 `bean-searcher.params.convertor.date-target`(TODO)
+  * 新增配置项 `bean-searcher.params.convertor.date-time-target`(TODO)
+  * 新增配置项 `bean-searcher.params.convertor.time-target`(TODO)
+* Bean Searcher Solon Plugin
+  * 新增配置项 `bean-searcher.params.fail-on-error` 指定参数错误时，是否抛出异常，默认 `false`(TODO)
+  * 新增配置项 `bean-searcher.params.convertor.date-target`(TODO)
+  * 新增配置项 `bean-searcher.params.convertor.date-time-target`(TODO)
+  * 新增配置项 `bean-searcher.params.convertor.time-target`(TODO)
 
 # v4.2.2 @ 2023-07-18